refactor: 모델 특성에 따라 하위 패키지 분류
