+++ src/main/java/blackjack/model/card/CardShape.java
@@ -1,4 +1,4 @@
-package blackjack.model;
+package blackjack.model.card;
 
 public enum CardShape {
     SPADE("스페이드"),
