+++ src/main/java/blackjack/model/game/ParticipantResult.java
@@ -1,4 +1,7 @@
-package blackjack.model;
+package blackjack.model.game;
+
+import blackjack.model.player.Dealer;
+import blackjack.model.player.Participant;
 
 public enum ParticipantResult {
     WIN,
@@ -25,7 +28,6 @@ public enum ParticipantResult {
         }
         return DRAW;
     }
-
 
 
 }
