+++ src/main/java/blackjack/model/player/Dealer.java
@@ -1,4 +1,4 @@
-package blackjack.model;
+package blackjack.model.player;
 
 public class Dealer extends Player {
