+++ src/main/java/blackjack/model/player/Participants.java
@@ -1,4 +1,4 @@
-package blackjack.model;
+package blackjack.model.player;
 
 import java.util.ArrayList;
 import java.util.HashSet;
