+++ src/main/java/blackjack/view/OutputView.java
@@ -1,12 +1,12 @@
 package blackjack.view;
 
-import blackjack.model.Card;
-import blackjack.model.Dealer;
-import blackjack.model.Participant;
-import blackjack.model.ParticipantResult;
-import blackjack.model.Participants;
-import blackjack.model.Player;
-import blackjack.model.ReceivedCards;
+import blackjack.model.card.Card;
+import blackjack.model.game.ParticipantResult;
+import blackjack.model.game.ReceivedCards;
+import blackjack.model.player.Dealer;
+import blackjack.model.player.Participant;
+import blackjack.model.player.Participants;
+import blackjack.model.player.Player;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.stream.Collectors;
