+++ src/test/java/blackjack/model/card/CardTest.java
@@ -1,4 +1,4 @@
-package blackjack.model;
+package blackjack.model.card;
 
 import static org.assertj.core.api.Assertions.assertThat;
