+++ src/test/java/blackjack/model/game/BlackJackGameTest.java
@@ -1,7 +1,13 @@
-package blackjack.model;
+package blackjack.model.game;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import blackjack.model.card.Card;
+import blackjack.model.card.CardShape;
+import blackjack.model.card.CardType;
+import blackjack.model.player.Dealer;
+import blackjack.model.player.Participant;
+import blackjack.model.player.Participants;
 import java.util.List;
 import org.junit.jupiter.api.Test;
