+++ src/test/java/blackjack/model/game/DeckInitializerTest.java
@@ -1,8 +1,7 @@
-package blackjack.model;
+package blackjack.model.game;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import java.util.List;
 import org.junit.jupiter.api.Test;
 
 class DeckInitializerTest {
