+++ src/test/java/blackjack/model/game/ParticipantResultTest.java
@@ -1,7 +1,12 @@
-package blackjack.model;
+package blackjack.model.game;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import blackjack.model.card.Card;
+import blackjack.model.card.CardShape;
+import blackjack.model.card.CardType;
+import blackjack.model.player.Dealer;
+import blackjack.model.player.Participant;
 import org.junit.jupiter.api.Test;
 
 class ParticipantResultTest {
