+++ src/test/java/blackjack/model/player/DealerTest.java
@@ -1,7 +1,10 @@
-package blackjack.model;
+package blackjack.model.player;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import blackjack.model.card.Card;
+import blackjack.model.card.CardShape;
+import blackjack.model.card.CardType;
 import org.junit.jupiter.api.Test;
 
 class DealerTest {
