+++ src/test/java/blackjack/model/player/ParticipantsTest.java
@@ -1,4 +1,4 @@
-package blackjack.model;
+package blackjack.model.player;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
